refactor: Rename file `user` to `object`
- so we can aggregate other objects together
